requirements.txt
livekit-agents>=0.9.1
livekit-plugins-openai>=0.8.5
livekit-plugins-deepgram>=0.6.7
livekit-plugins-silero>=0.6.4
python-dotenv~=1.0
